Begin economy integration, fix several bugs
Been extremely busy lately and have not been able to work on this as
much as I'd like to, that being said I was able to further optimize the
tab function (tested with 300 moving bots, 20tps - before was 4tps) and
also begin adding economy functions. Plugin is now dependent on Vault,
which has been added to the pom.xml.